requirements.txt
Poutyne==0.5
seaborn==0.9.0
numpy==1.15.0
torch
scikit_image==0.14.1
scipy==1.1.0
matplotlib==3.0.2
torchvision
Pillow==6.0.0
scikit_learn==0.20.3
